chores: move dto schema to other dedicated file
